simplify code + remove duplicate code, update image segmentation experiments
