+++ src/image_segmentation/fcn_img_segmentation.py
@@ -1,11 +1,10 @@
-from __future__ import print_function, division
 import argparse
 import time
 import matplotlib.pyplot as plt
 import numpy as np
 import torch
 import torch.nn.functional as F
-import torch.optim as optim
+from torch.utils.data import DataLoader
 from torchsummary import summary
 import os
 
@@ -14,224 +13,188 @@ from src.image_segmentation.weizmann_horse_dataset import (
 )
 from src.image_segmentation.utils import average_over_crops, show_preds_test_time, get_iou_batch
 from src.image_segmentation.model.fcn import FCN
+from src.model.base_model import BaseModel
+from src.visualization_utils import plot_results
 
 __author__ = "HSU CHIH-CHAO and Philippe Beardsell. University of Montreal"
 
 FILE_SAVE_FCN = "fcn_best.pth"
 
 
-def train(model, device, train_loader, optimizer):
-    model.train()
-    train_loss = 0
-    total_iou = 0
-    # define the operation batch-wise
-    for batch_idx, (raw_inputs, data, target) in enumerate(train_loader):
-        # send the data into GPU or CPU
-        data, target = data.to(device), target.to(device)
-        target[target > 0] = 1
+class FCNModel(BaseModel):
 
-        # clear the gradient in the optimizer in the begining of each backpropagation
-        optimizer.zero_grad()
+    def __init__(self, lr: float, momentum: float):
+        super().__init__(model=FCN())
+        self.optimizer = torch.optim.SGD(self.model.parameters(), lr=lr, momentum=momentum)
 
-        output = model(data)
+    @staticmethod
+    def _compute_pred_and_loss(output, target):
+        n, c, h, w = output.size()
+        log_p = F.log_softmax(output, dim=1)
+        # prediction (pick higher probability after log softmax)
+        pred = torch.argmax(log_p, dim=1)
+        # log_p: (n*h*w, c)
+        log_p = log_p.transpose(1, 2).transpose(2, 3)
+        target = target.transpose(1, 2).transpose(2, 3)
+        log_p = log_p[target.repeat(1, 1, 1, c) >= 0]
+        log_p = log_p.view(-1, c)
+        # target: (n*h*w,)
+        mask = target >= 0
+        target = target[mask]
 
-        pred, loss = compute_pred_and_loss(output, target)
+        loss = F.nll_loss(log_p, target.long())
+        return pred, loss
 
-        total_iou += torch.mean(get_iou_batch(pred.detach(), target.long()))
+    def train(self, loader: DataLoader):
+        self.model.train()
+        train_loss = 0
+        total_iou = 0
+        # define the operation batch-wise
+        for inputs, targets in loader:
+            inputs, targets = inputs.to(self.device), targets.to(self.device)
+            targets[targets > 0] = 1
 
-        loss.backward()
-        # update the parameters
-        optimizer.step()
-        # train loss
-        train_loss += loss.item()
+            # clear the gradient in the optimizer in the begining of each backpropagation
+            self.optimizer.zero_grad()
 
-    return train_loss/len(train_loader), total_iou/len(train_loader)
+            output = self.model(inputs)
 
+            pred, loss = self._compute_pred_and_loss(output, targets)
 
-def compute_pred_and_loss(output, target):
-    n, c, h, w = output.size()
-    log_p = F.log_softmax(output, dim=1)
-    # prediction (pick higher probability after log softmax)
-    pred = torch.argmax(log_p, dim=1)
-    # log_p: (n*h*w, c)
-    log_p = log_p.transpose(1, 2).transpose(2, 3)
-    target = target.transpose(1, 2).transpose(2, 3)
-    log_p = log_p[target.repeat(1, 1, 1, c) >= 0]
-    log_p = log_p.view(-1, c)
-    # target: (n*h*w,)
-    mask = target >= 0
-    target = target[mask]
+            total_iou += torch.mean(get_iou_batch(pred.detach(), targets.long()))
 
-    loss = F.nll_loss(log_p, target.long())
+            loss.backward()
+            # update the parameters
+            self.optimizer.step()
+            # train loss
+            train_loss += loss.item()
 
-    return pred, loss
+        return train_loss / len(loader), total_iou / len(loader)
 
+    def valid(self, loader: DataLoader):
+        self.model.eval()
+        test_loss = 0
+        total_iou = 0
 
-def visualize(iou, data, target, pred):
-    i = iou.max(0)[1].item()
-    plt.figure()
-    print(iou[i])
-    plt.imshow(data[i].to('cpu').numpy().transpose((1, 2, 0)))
-    plt.figure()
-    plt.imshow(np.squeeze(target[i].to('cpu').numpy().transpose((1, 2, 0))))
-    plt.figure()
-    plt.imshow(pred[i].to('cpu').numpy())
+        with torch.no_grad():
+            for inputs, targets in loader:
+                inputs, targets = inputs.to(self.device), targets.to(self.device)
 
+                targets[targets > 0] = 1
+                output = self.model(inputs)
 
-def valid(model, device, test_loader):
-    model.eval()
-    test_loss = 0
-    total_iou = 0
+                pred, loss = self._compute_pred_and_loss(output, targets)
 
-    with torch.no_grad():
-        for (raw_inputs, data, target) in test_loader:
-            data, target = data.to(device), target.to(device)
+                iou = get_iou_batch(pred, targets.long())
+                total_iou += torch.mean(iou)
 
-            target[target > 0] = 1
-            output = model(data)
+                # if epochs % 25== 0:
+                #    visualize(iou, inputs, pred, targets)
 
-            pred, loss = compute_pred_and_loss(output, target)
+                # Average the loss (batch_wise)
+                test_loss += loss.item()
 
-            iou = get_iou_batch(pred, target.long())
-            total_iou += torch.mean(iou)
+        return test_loss / len(loader), total_iou / len(loader)
 
-            # if epochs % 25== 0:
-            #    visualize(iou, data, pred, target)
+    def test(self, loader: DataLoader, show_n_samples: int = 0):
+        """
+            At Test time, we are averaging our predictions
+            over 36 crops of 24x24 mask to predict a 32x32 mask
+        """
+        self.model.eval()
+        mean_iou = []
+        n_shown = 0
 
-            # Average the loss (batch_wise)
-            test_loss += loss.item()
+        with torch.no_grad():
+            for raw_inputs, inputs, targets in loader:
+                inputs, targets = inputs.to(self.device), targets.to(self.device)
 
-    return test_loss/len(test_loader), total_iou/len(test_loader)
+                # For test: inputs is a 5d tensor
+                bs, ncrops, channels, h, w = inputs.size()
 
+                # fuse batch size and ncrops to know our estimated IOU
+                output = self.model(inputs.view(-1, channels, h, w))
+                log_p = F.log_softmax(output, dim=1)
+                pred = torch.argmax(log_p, dim=1)
+                # go back to normal shape and take the mean in the 1 dim
+                pred = pred.view(bs, ncrops, h, w)
+                # output = output.view(bs, ncrops, h, w)
 
-def test(model, loader, device):
-    """
-    At Test time, we are averaging our predictions
-    over 36 crops of 24x24 mask to predict a 32x32 mask
-    """
-    model.eval()
-    mean_iou = []
+                final_pred = average_over_crops(pred, self.device)
+                oracle = get_iou_batch(final_pred, targets.float())
+                for o in oracle:
+                    mean_iou.append(o)
 
-    with torch.no_grad():
-        for batch_idx, (raw_inputs, inputs, targets) in enumerate(loader):
-            inputs, targets = inputs.to(device), targets.to(device)
+                if n_shown < show_n_samples:
+                    n_shown += 1
+                    show_preds_test_time(raw_inputs, final_pred)
 
-            # For test: inputs is a 5d tensor
-            bs, ncrops, channels, h, w = inputs.size()
+        mean_iou = torch.stack(mean_iou)
+        mean_iou = torch.mean(mean_iou)
+        mean_iou = mean_iou.cpu().numpy()
 
-            # fuse batch size and ncrops to know our estimated IOU
-            output = model(inputs.view(-1, channels, h, w))
-            log_p = F.log_softmax(output, dim=1)
-            pred = torch.argmax(log_p, dim=1)
-            # go back to normal shape and take the mean in the 1 dim
-            pred = pred.view(bs, ncrops, h, w)
-            #                output = output.view(bs, ncrops, h, w)
+        print('Test set: IOU = {:.2f}%'.format(100 * mean_iou))
+        return mean_iou
 
-            final_pred = average_over_crops(pred, device)
-            oracle = get_iou_batch(final_pred, targets.float())
-            for o in oracle:
-                mean_iou.append(o)
-
-            show_preds_test_time(raw_inputs, final_pred, oracle)
-
-    mean_iou = torch.stack(mean_iou)
-    mean_iou = torch.mean(mean_iou)
-    mean_iou = mean_iou.cpu().numpy()
-
-    print('Test set: IOU = {:.2f}%'.format(100 * mean_iou))
-
-    return mean_iou
-
-
-def plot_loss_iou(n_epochs: int, train_loss, valid_loss, train_iou, valid_iou) -> None:
-    x = list(range(1, n_epochs + 1))
-    # plot train/validation loss versus epoch
-    plt.figure()
-    plt.title("Train/Validation Loss")
-    plt.xlabel("Epochs")
-    plt.ylabel("Total Loss")
-    plt.plot(x, train_loss, label="train loss")
-    plt.plot(x, valid_loss, color='red', label="validation loss")
-    plt.legend(loc='upper right')
-    plt.grid(True)
-    plt.show()
-
-    # plot train/validation loss versus epoch
-    plt.figure()
-    plt.title("Train/Validation IOU")
-    plt.xlabel("Epochs")
-    plt.ylabel("Mean IOU")
-    plt.plot(x, train_iou, label="train iou")
-    plt.plot(x, valid_iou, color='red', label="validation iou")
-    plt.legend(loc='upper right')
-    plt.grid(True)
-    plt.show()
+    @staticmethod
+    def _visualize(iou, inputs, targets, pred):
+        i = iou.max(0)[1].item()
+        plt.figure()
+        plt.imshow(inputs[i].to('cpu').numpy().transpose((1, 2, 0)))
+        plt.figure()
+        plt.imshow(np.squeeze(targets[i].to('cpu').numpy().transpose((1, 2, 0))))
+        plt.figure()
+        plt.imshow(pred[i].to('cpu').numpy())
 
 
 def main():
-    # Version of Pytorch
-    print("Pytorch Version:", torch.__version__)
-
     # Training args
     parser = argparse.ArgumentParser(description='Fully Convolutional Network')
-    parser.add_argument('--batch-size', type=int, default=32, metavar='N',
-                        help='input batch size for training (default: 64)')
-    parser.add_argument('--test-batch-size', type=int, default=128, metavar='N',
-                        help='input batch size for testing (default: 1000)')
-    parser.add_argument('--epochs', type=int, default=200, metavar='N',
-                        help='number of epochs to train (default: 10)')
+    parser.add_argument('--batch-size', type=int, default=64, metavar='N',
+                        help='input batch size for training')
+    parser.add_argument('--test-batch-size', type=int, default=1000, metavar='N',
+                        help='input batch size for testing')
+    parser.add_argument('--epochs', type=int, default=100, metavar='N',
+                        help='number of epochs to train')
     parser.add_argument('--lr', type=float, default=0.1, metavar='LR',
-                        help='learning rate (default: 0.01)')
+                        help='learning rate')
     parser.add_argument('--momentum', type=float, default=0.9, metavar='M',
-                        help='SGD momentum (default: 0.5)')
-    parser.add_argument('--no-cuda', action='store_true', default=True,
-                        help='disables CUDA training')
-    parser.add_argument('--seed', type=int, default=1, metavar='S',
-                        help='random seed (default: 1)')
-    parser.add_argument('--log-interval', type=int, default=1000, metavar='N',
-                        help='how many batches to wait before logging training status')
+                        help='SGD momentum')
 
-    parser.add_argument('--save-model', action='store_true', default=True,
-                        help='For Saving the current Model')
     args = parser.parse_args()
 
     path_save_model = os.path.join(PATH_SAVE_HORSE, FILE_SAVE_FCN)
-    n_epochs = args.epochs
     # Use GPU if it is available
     use_cuda = torch.cuda.is_available()
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+    device = torch.device("cuda" if use_cuda else "cpu")
 
     # Create FCN
-    model = FCN().to(device)
+    fcn = FCNModel(args.lr, args.momentum)
     # print the model summary
-    print(model)
+    print(fcn.model)
 
     # Visualize the output of each layer via torchSummary
-    summary(model, input_size=(3, 24, 24))
+    summary(fcn.model, input_size=(3, 24, 24))
 
     train_loader, valid_loader = load_train_set_horse(PATH_DATA_WEIZMANN, use_cuda,
                                                       args.batch_size, args.batch_size)
 
-    optimizer = optim.SGD(model.parameters(), lr=args.lr, momentum=args.momentum)
-
-    train_loss = []
-    train_iou = []
-    validation_loss = []
-    validation_iou = []
+    train_loss, validation_loss = [], []
+    train_iou, validation_iou = [], []
     best_val_iou = 0
 
     start_time = time.time()
     # Start Training
-    for epoch in range(1, n_epochs + 1):
+    for epoch in range(1, args.epochs + 1):
         # train loss
-        t_loss, t_mean_iou = train(model, device, train_loader, optimizer)
-        print('Train Epoch: {} \t Loss: {:.6f}\t Mean_IOU(%):{}%'.format(
-            epoch, t_loss, t_mean_iou))
+        t_loss, t_mean_iou = fcn.train(train_loader)
+        print('Train Epoch: {} \t Loss: {:.6f}\t Mean_IOU: {:.2f}%'.format(
+            epoch, t_loss, 100 * t_mean_iou))
 
         # validation loss
-        v_loss, v_mean_iou = valid(model, device, valid_loader)
-        print('Validation Epoch: {} \t Loss: {:.6f}\t Mean_IOU(%):{}%'.format(
-            epoch, v_loss, v_mean_iou))
+        v_loss, v_mean_iou = fcn.valid(valid_loader)
+        print('Validation Epoch: {} \t Loss: {:.6f}\t Mean_IOU: {:.2f}%'.format(
+            epoch, v_loss, 100 * v_mean_iou))
 
         train_loss.append(t_loss)
         train_iou.append(t_mean_iou)
@@ -239,24 +202,20 @@ def main():
         validation_iou.append(v_mean_iou)
         if v_mean_iou > best_val_iou:
             best_val_iou = v_mean_iou
-            print('--- Saving model at IOU_{:.2f} ---'.format(100 * best_val_iou))
-            torch.save(model.state_dict(), path_save_model)
+            print('--- Saving model at IOU: {:.2f}% ---'.format(100 * best_val_iou))
+            torch.save(fcn.model.state_dict(), path_save_model)
 
         print('-------------------------------------------------------')
 
-    print("--- %s seconds ---" % (time.time() - start_time))
+    print("--- {:.2f} seconds ---".format(time.time() - start_time))
 
     # Visualization
-    plot_loss_iou(n_epochs, train_loss, validation_loss,
-                  train_iou, validation_iou)
+    plot_results("IOU", train_loss, validation_loss, validation_iou, train_iou)
 
-    # test set
-    print("Best Validation Mean IOU:", best_val_iou)
+    print("Best Validation Mean IOU: {:.2f}%".format(100 * best_val_iou))
 
     # Test on 36 crop
-    fcn_test = FCN().to(device)
-    fcn_test.load_state_dict(torch.load(path_save_model))
-    fcn_test.eval()
+    fcn.model.load_state_dict(torch.load(path_save_model, map_location=device))
 
     # Compute IOU single prediction on 24x24 crops and 36 crops averaging on 32x32
     for i in range(2):
@@ -267,16 +226,12 @@ def main():
 
         print('-------------------------------------------')
         if i == 0:
-            mean_iou = 0
             print('Single crop IOU prediction')
-            for epoch in range(1, 100):
-                # validation loss
-                v_loss, v_mean_iou = valid(model, device, valid_loader)
-                mean_iou += v_mean_iou
-            print('Validation Mean_IOU(%):{}%'.format(mean_iou / 100))
+            loss, mean_iou = fcn.valid(valid_loader)
+            print('Validation Mean_IOU: {:.2f}%'.format(mean_iou * 100))
         else:
             print('36 Crops IOU prediction')
-            test(fcn_test, test_loader, device)
+            fcn.test(test_loader, show_n_samples=1)
 
 
 if __name__ == "__main__":
